Updated to 1.04
Note:  For Chams, weapons_flag, dead_bodies_flag, they seemed to have
added a extra parameter for the functions parameters you were editing
for those mods. So I updated it so that the mods get written to the new
register that they added. I did not get a chance to test, so you'll have
to do that. Though all the other mods should work fine. If the 3 I said
don't then look in IDA to see what I mean.